% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{closeDb}
\alias{closeDb}
\title{Finalise and close the database}
\usage{
closeDb(db, experimentSet = "")
}
\arguments{
\item{db}{Database object created by \code{\link{initializeDb}}}

\item{experimentSet}{Name of the experiment set for all the experiments (optional)}
}
\description{
Finalise and close the database
}
\examples{
  db <- initializeDb("", "My Dataset")
  # ...
  closeDb(db)
}

