% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{insertAliases}
\alias{insertAliases}
\title{Insert aliases}
\usage{
insertAliases(db, aliasType, aliases, aliasSource = "",
  aliasDescription = "")
}
\arguments{
\item{db}{Database object created by \code{\link{initializeDb}}}

\item{aliasType}{Name of the alias}

\item{aliases}{A vector of aliases with names giving the primary identifiers}

\item{aliasSource}{Optional alias metadata: source}

\item{aliasDescription}{Optional alias metadata: description}
}
\value{
An updated database object db
}
\description{
Insert aliases
}
\examples{
  db <- initializeDb("", "My Dataset")
  aliases <- c("aliasA", "aliasB", "aliasC")
  names(aliases) <- c("A", "B", "C")
  db <- insertAliases(db, "testalias", aliases)
  closeDb(db)
}

