% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{insertZScores}
\alias{insertZScores}
\title{Insert data z-scores used for filtering}
\usage{
insertZScores(db, zscores)
}
\arguments{
\item{db}{Database object created by \code{\link{initializeDb}}}

\item{zscores}{A vector of z-scores of elements identified by names}
}
\value{
An updated database object db
}
\description{
Insert data z-scores used for filtering
}
\examples{
  db <- initializeDb("", "My Dataset")
  zscores <- c(1, 2, 3)
  names(zscores) <- c("A", "B", "C")
  db <- insertZScores(db, zscores)
  closeDb(db)
}

